requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
PyMySQL==1.1.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
jieba==0.42.1
numpy==1.24.3
requests==2.31.0
